requirements.txt
torchtext==0.3.1
spacy

astroid==1.6.0
click==6.7
Flask==0.12.2
gunicorn==19.7.1
isort==4.2.15
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pylint==1.8.1
six==1.11.0
Werkzeug==0.14.1
wrapt==1.10.11

redis==3.2.0
celery==4.3.0
pyyaml
kubernetes